test(web): replace Labs mocks with composed page journeys
